# new_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new_server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in new_server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "requestId"],
        "properties": {
          "error": { "type": "string" },
          "details": {},
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      },
...
      "GrantedBy": {
        "oneOf": [
          {
            "type": "object",
            "required": ["userId", "username", "displayName"],
            "properties": {
              "userId": { "type": "integer" },
              "username": { "type": "string" },
              "displayName": { "type": "string", "nullable": true }
            }
          },
          {
            "type": "object",
            "required": ["label"],
            "properties": {
              "label": { "type": "string" }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "PermissionUser": {
        "type": "object",
        "required": ["id", "username", "displayName", "telegramId", "status"],
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "telegramId": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "disabled"] }
        }
      },
      "Permission": {
        "type": "object",
        "required": ["id", "domain", "localPart", "emailAddress", "role", "status", "user", "grantedBy", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "domain": { "type": "string" },
          "localPart": { "type": "string", "nullable": true },
          "emailAddress": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "status": { "type": "string", "enum": ["active", "disabled"] },
          "user": { "$ref": "#/components/schemas/PermissionUser" },
          "grantedBy": { "$ref": "#/components/schemas/GrantedBy" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "PermissionCreateRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "userId": { "oneOf": [{ "type": "integer" }, { "type": "string" }] },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "telegramId": { "type": "string", "nullable": true },
          "domain": { "type": "string" },
          "localPart": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "status": { "type": "string", "enum": ["active", "disabled"] }
        }
      },
      "PermissionUpdateRequest": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "status": { "type": "string", "enum": ["active", "disabled"] }
        }
      },
...
      "SimpleSuccessResponse": {
        "type": "object",
        "required": ["success", "requestId"],
        "properties": {
          "success": { "type": "boolean" },
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/permissions": {
      "get": {
        "tags": ["Permissions"],
        "summary": "List permissions",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "query", "name": "userId", "schema": { "type": "string" } },
          { "in": "query", "name": "username", "schema": { "type": "string" } },
          { "in": "query", "name": "domain", "schema": { "type": "string" } },
          { "in": "query", "name": "localPart", "schema": { "type": "string" } },
          { "in": "query", "name": "role", "schema": { "type": "string", "enum": ["viewer", "operator", "admin"] } },
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["active", "disabled"] } }
        ],
        "responses": {
          "200": {
            "description": "Permission list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "permissions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Permission" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Permissions"],
        "summary": "Create or upsert permission",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PermissionCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/permissions/{permissionId}": {
      "get": {
        "tags": ["Permissions"],
        "summary": "Get permission by ID",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Permission details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Permissions"],
        "summary": "Update permission",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PermissionUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Permissions"],
        "summary": "Delete permission",
        "description": "Global admin only",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/v1/domains/{domain}/permissions": {
      "get": {
        "tags": ["Domains"],
        "summary": "List permissions in a domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Domain permission list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "permissions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Permission" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Domains"],
        "summary": "Create or upsert domain permission",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  { "$ref": "#/components/schemas/PermissionCreateRequest" },
                  {
                    "type": "object",
                    "properties": {
                      "domain": { "readOnly": true }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/domains/{domain}/permissions/{permissionId}": {
      "get": {
        "tags": ["Domains"],
        "summary": "Get permission in domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Permission details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Domains"],
        "summary": "Update permission in domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PermissionUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "permission": { "$ref": "#/components/schemas/Permission" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Domains"],
        "summary": "Delete permission in domain",
        "description": "Requires domain admin or global admin",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "domain", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "permissionId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}